+++ src/xddmod/src/handlers.rs
@@ -1,3 +1,3 @@
-pub mod gamba;
+pub mod gambage;
 pub mod npc;
 pub mod persistence;
